+++ src/icemac/addressbook/conftest.py
@@ -12,10 +12,7 @@ import icemac.addressbook.principals.principals
 import icemac.addressbook.startup
 import icemac.addressbook.testing
 import icemac.addressbook.utils
-import os
-import os.path
 import pytest
-import tempfile
 import transaction
 import zope.app.wsgi.testlayer
 import zope.browserpage.metaconfigure
@@ -81,30 +78,6 @@ def browser2(browserWsgiAppS):
 def browser3(browserWsgiAppS):
     """Fixture for testing with a third zope.testbrowser."""
     return icemac.addressbook.testing.Browser(wsgi_app=browserWsgiAppS)
-
-
-@pytest.yield_fixture(scope='function')
-def tmpfile():
-    """Fixture to create a temporary file with a defined content and suffix.
-
-    Returns a callable to create the file which returns a tuple of file handle
-    and file name.
-
-    """
-    data = {'filename': None, 'fhr': None}
-
-    def tmpfile(content, suffix, _data=data):
-        fd, filename = tempfile.mkstemp(suffix=suffix)
-        # It does not seem to be possible to use os.fdopen with 'rw' -- it
-        # leads to "[Errno 9] Bad file descriptor"
-        with os.fdopen(fd, 'w') as fhw:
-            fhw.write(content)
-        data['fhr'] = file(filename, 'r')
-        data['filename'] = filename
-        return data['fhr'], os.path.basename(filename)
-    yield tmpfile
-    data['fhr'].close()
-    os.unlink(data['filename'])
 
 
 # Fixtures to help asserting